requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.5.0
Requests==2.32.3
schedule==1.2.0
